+++ packages/alex/src/api/renderApp.tsx
@@ -1,10 +1,12 @@
-import React, { useState, useMemo, useEffect } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import ReactDOM from 'react-dom';
-import { ClientApp } from '@alipay/alex-core';
+import { IReporterService, localize, getDebugLogger } from '@ali/ide-core-common';
+import { REPORT_NAME } from '@alipay/alex-core';
 import { createApp } from './createApp';
 import { Root } from '../core/Root';
 import { RootProps, LandingProps } from '../core/types';
 import { themeStorage } from '../core/utils';
+import { useConstant } from '../core/hooks';
 import { IConfig, IAppInstance } from './types';
 
 export interface IRenderProps extends IConfig {
@@ -34,11 +36,18 @@ export const renderApp = (domElement: HTMLElement, props: IRenderProps) => {
       onLoad?.(app);
     })
     .catch((err: Error) => {
-      console.error(err);
       ReactDOM.render(
-        <Root status="error" error={err?.message || ''} theme={themeType} />,
+        <Root status="error" error={err?.message || localize('error.unknown')} theme={themeType} />,
         domElement
       );
+
+      (app.injector.get(
+        IReporterService
+      ) as IReporterService).point(REPORT_NAME.ALEX_APP_START_ERROR, err?.message, { error: err });
+      getDebugLogger().error(err);
+      setTimeout(() => {
+        throw err;
+      });
     });
 
   return () => {
@@ -47,8 +56,9 @@ export const renderApp = (domElement: HTMLElement, props: IRenderProps) => {
 };
 
 export const AppRenderer: React.FC<IRenderProps> = ({ onLoad, Landing, ...opts }) => {
-  const themeType = useMemo(() => themeStorage.get(), []);
-  const [appElement, setAppElement] = useState<React.ReactElement | null>(null);
+  const app = useConstant(() => createApp(opts));
+  const themeType = useConstant(() => themeStorage.get());
+  const appElementRef = useRef<React.ReactElement | null>(null);
 
   const [state, setState] = useState<{
     status: RootProps['status'];
@@ -56,19 +66,27 @@ export const AppRenderer: React.FC<IRenderProps> = ({ onLoad, Landing, ...opts }
   }>(() => ({ status: 'loading' }));
 
   useEffect(() => {
-    const app = createApp(opts);
     app
       .start((appElement) => {
-        setAppElement(appElement);
+        appElementRef.current = appElement;
+        setState({ status: 'success' });
         return Promise.resolve();
       })
       .then(() => {
-        setState({ status: 'success' });
         onLoad?.(app);
       })
       .catch((err: Error) => {
-        console.error(err);
-        setState({ error: err?.message || '', status: 'error' });
+        setState({ error: err?.message || localize('error.unknown'), status: 'error' });
+
+        (app.injector.get(
+          IReporterService
+        ) as IReporterService).point(REPORT_NAME.ALEX_APP_START_ERROR, err?.message, {
+          error: err,
+        });
+        getDebugLogger().error(err);
+        setTimeout(() => {
+          throw err;
+        });
       });
 
     return () => {
@@ -78,7 +96,7 @@ export const AppRenderer: React.FC<IRenderProps> = ({ onLoad, Landing, ...opts }
 
   return (
     <Root {...state} theme={themeType} Landing={Landing}>
-      {appElement}
+      {appElementRef.current}
     </Root>
   );
 };