+++ packages/alex/src/integration/startup/index.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { GitFileSchemeModule } from '@alipay/alex-git';
-import { AppRenderer, renderApp } from '../../api/renderApp';
+import { IAppInstance, AppRenderer } from '../..';
+import { GitFileSchemeModule } from '@alipay/alex-code-service';
+import * as Alex from '../..';
 import { StartupModule } from './startup.module';
 import './languages';
 import SarifViewer from '../../../extensions/cloud-ide-ext.sarif-viewer';
@@ -10,6 +11,8 @@ import html from '../../../extensions/alex.html-language-features-worker';
 import json from '../../../extensions/alex.json-language-features-worker';
 import markdown from '../../../extensions/alex.markdown-language-features-worker';
 import typescript from '../../../extensions/alex.typescript-language-features-worker';
+
+(window as any).alex = Alex;
 
 const query = location.search
   .slice(1)
@@ -25,13 +28,7 @@ const project = query.project || 'ide-s/TypeScript-Node-Starter';
 ReactDOM.render(
   <AppRenderer
     onLoad={(app) => {
-      // (window as any).app = app
-      // app.service.fs.onFilesChanged(e => {
-      //   e.forEach(async item => {
-      //     console.log('>>>content', await app.service.fs.resolveContent(item.uri))
-      //     console.log('>>>path', item.uri)
-      //   })
-      // })
+      window.app = app;
     }}
     appConfig={{
       modules: [GitFileSchemeModule, StartupModule],
@@ -39,11 +36,15 @@ ReactDOM.render(
       workspaceDir: project,
     }}
     runtimeConfig={{
-      git: {
+      codeService: {
+        platform: 'antcode',
         baseURL: '/code-service',
         project,
         branch: query.branch,
         commit: query.commit,
+        transformStaticResource({ baseURL, project, commit, path }) {
+          return `${baseURL}/${project}/raw/${commit}/${path}`;
+        },
       },
       // workspace: {
       //   filesystem: {
@@ -65,22 +66,14 @@ ReactDOM.render(
   document.getElementById('main')
 );
 
-// renderApp(document.getElementById('main')!, {
-//   appConfig: {
-//     modules: [StartupModule]
-//   },
-//   runtimeConfig: {
-//     // TODO: 从 query 上取
-//     git: {
-//       baseURL: '/code-service',
-//       project: query.project || 'ide-s/TypeScript-Node-Starter',
-//       branch: query.branch,
-//       commit: query.commit,
-//     },
-//   },
-// });
-
 // for test
-(window as any).destroy = () => {
+window.destroy = () => {
   ReactDOM.render(<div>destroyed</div>, document.getElementById('main'));
 };
+
+declare global {
+  interface Window {
+    app: IAppInstance;
+    destroy(): void;
+  }
+}