+++ packages/alex/src/modules/ali__ide-core-browser.ts
@@ -1,1 +1,8 @@
+import { IToolbarActionGroup } from '@ali/ide-core-browser';
+
 export * from '@ali/ide-core-browser';
+export * from '@ali/ide-core-browser/lib/menu/next';
+export * from '@ali/ide-core-browser/lib/context-key';
+
+// local exports take priority.
+export { IToolbarActionGroup };