+++ packages/core/src/common/constant.ts
@@ -10,7 +10,9 @@ export const TMP_ROOT = os.tmpdir();
 
 export const WORKSPACE_ROOT = '/workspace';
 
-export const GIT_ROOT = '/git';
+export const CODE_ROOT = '/code';
+
+export const IDB_ROOT = '/idb';
 
 // 全局数据存储目录
 export const STORAGE_NAME = '.cloudide';