+++ packages/core/src/server/core/util.ts
@@ -6,8 +6,7 @@ const { createFileSystem, FileSystem, initialize } = BrowserFS;
 
 export const initializeRootFileSystem = async () => {
   // TODO: /home 采用 localStorage 作为文件系统是不是更好点，理论数据量不大的话没必要用 IndexedDB
-  const [tmp, workspace, home] = await Promise.all([
-    createFileSystem(FileSystem.InMemory, {}),
+  const [tmp, home] = await Promise.all([
     createFileSystem(FileSystem.InMemory, {}),
     createFileSystem(FileSystem.IndexedDB, {
       storeName: HOME_IDB_NAME,
@@ -16,7 +15,6 @@ export const initializeRootFileSystem = async () => {
   const mountfs = await createFileSystem(FileSystem.MountableFileSystem, {
     [TMP_ROOT]: tmp,
     [HOME_ROOT]: home,
-    [WORKSPACE_ROOT]: workspace,
   });
   initialize(mountfs);
   return mountfs as RootFS;